requirements.txt
anndata==0.10.8
arrow==1.3.0
h5py==3.11.0
jupyter==1.1.1
jupyterlab==4.2.5
lightning==2.1.4
lightning-bolts @ git+https://github.com/PytorchLightning/lightning-bolts.git@2c4602aa684e7b90e7ffdcea1d3f93a20f9c2ead
lightning-lite==1.8.0
lightning-utilities==0.11.6
llvmlite==0.43.0
loompy==3.0.7
matplotlib==3.9.1
matplotlib-inline==0.1.7
notebook==7.2.2
numba==0.60.0
numpy==1.26.3
opencv-python==4.10.0.84
pandas==2.2.3
pytorch-lightning==2.4.0
rpds-py==0.20.0
scanpy==1.10.2
scikit-image==0.24.0
scikit-learn==1.5.1
scikit-misc==0.5.1
scipy==1.14.0
scvi-tools==1.1.5
seaborn==0.13.2
statsmodels==0.14.2
stdlib-list==0.10.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tifffile==2024.7.24
torch==2.4.0+cu124
torchmetrics==0.11.4
torchvision==0.19.0+cu124
tqdm==4.66.5

